refactor: #296 - 노트 RAG 검색 로직 공통화
