+++ src/features/ai/rags/note/constants/embeddings.ts
